pyproject.toml
[project]
name = "deepseek-ocr-mcp"
version = "0.1.0"
description = "DeepSeek-OCR MCP Server powered by FastMCP"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
deepseek-ocr = "deepseek_ocr_server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"